Provided almoust final refactoring
