+++ backend/Origam.Architect.Server/Interfaces/Services/IScreenActions.cs
@@ -1,0 +1,38 @@
+#region license
+/*
+Copyright 2005 - 2026 Advantage Solutions, s. r. o.
+
+This file is part of ORIGAM (http://www.origam.org).
+
+ORIGAM is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+ORIGAM is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with ORIGAM. If not, see <http://www.gnu.org/licenses/>.
+*/
+#endregion
+
+using Origam.Architect.Server.Models.Requests.Actions;
+using Origam.Architect.Server.Models.Responses.Actions;
+
+namespace Origam.Architect.Server.Interfaces.Services;
+
+/// <summary>
+/// Create-Screen wizard: read the entity's columns and existing DataStructure
+/// names, then generate the DataStructure + Screen Section + Screen.
+/// </summary>
+public interface IScreenActions
+{
+    /// <summary>Entity columns and existing DataStructure names for the wizard.</summary>
+    ScreenWizardData GetScreenWizardData(Guid entityId);
+
+    /// <summary>Generate DataStructure + PanelControlSet + FormControlSet from the entity.</summary>
+    CreateActionResult CreateScreen(CreateScreenModel input);
+}
